First release
This commit introduces a first working version of the tool.
I originally developed it about two years ago, when I first time
discovered the Terraria game. It fetches the Terraria server
executable that is published by creators on the Terraria official
website (terraria.org) and then assists in launching the server
inside a Linux container.

Recently I updated the code for the latest Terraria server version
and I decided to make it public.